// src/app/modules/trainer/Model/feedback.ts
export class Feedback {
    id: number;
    createdAt: Date;
    suggestions: String;
    trainer:number;
    trainee:number;
    feedback1: String;
    feedback2: String;
    feedback3: String;
    feedback4: String;
    feedback5: String;
    feedback6: String;
    feedback7: String;
    feedback8: String;
    feedback9: String;
    feedback10: String;
    feedback11: String;
    feedback12: String;
    feedback13: String;
    feedback14: String;
    feedback15:String;
    rate1:number;
    rate2:number;
    rate3:number;
    rate4:number;
    rate5:number;
    rate6:number;
    rate7:number;
    rate8:number;
    rate9:number;
    rate10:number;
    rate11:number;
    rate12:number;
    rate13:number;
    rate14:number;
    rate15:number;

}